requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
asyncio-atexit==1.0.1
attrdict==2.0.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bce-python-sdk==0.9.11
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
dnspython==2.6.1
docx==0.2.4
dynaconf==3.2.5
ecdsa
email_validator==2.1.1
emoji==2.12.1
erniebot==0.5.5
erniebot-agent==0.5.2
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
future
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imgaug==0.4.0
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langdetect==1.0.9
langserve==0.2.1
langsmith==0.1.67
lazy-object-proxy==1.10.0
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
onnx==1.14.1
onnxruntime==1.16.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.82
openpyxl==3.1.3
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1.post120
pandas==2.2.2
parso==0.8.4
patch-ng==1.17.4
pathable==0.4.3
pdf2docx==0.5.8
pdfminer.six==20231228
pillow==10.3.0
platformdirs==4.2.2
premailer==3.10.0
prompt_toolkit==3.0.46
protobuf==3.20.2
psutil==5.9.8
pure-eval==0.2.2
pyasn1 
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodome
pydantic
pydantic_core==2.18.3
Pygments==2.18.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
rarfile==4.2
referencing==0.31.1
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.18.1
rsa 
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
tifffile==2024.5.22
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.3
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.30.0
visualdl==2.5.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
